refactor(github-actions-usage): Split data logic across files
Organise the logic into three distinct steps:
1. Read data
2. Transform data
3. Write data

This should improve readability.
